add edit profile page. working on linking profile page to data model
