+++ app/imports/startup/client/router.js
@@ -22,6 +22,13 @@ FlowRouter.route('/:username/profile', {
   },
 });
 
+// FlowRouter.route('/:username/edit-profile', {
+//   name: 'Edit_My_Profile_Page',
+//   action() {
+//     BlazeLayout.render('App_Body', { main: 'Edit_My_Profile_Page' });
+//   },
+// });
+
 FlowRouter.route('/:username/adventures', {
   name: 'Adventures_Page',
   action() {
