+++ app/imports/ui/components/form-controls/checkbox-field.html
@@ -1,0 +1,21 @@
+<template name="Checkbox_Field">
+  {{!
+  Helper function for Checkbox_Form_Control.
+
+  Parameters:
+    label, name, id, value (String)
+    checked (boolean)
+
+  Label is required. name, id, and value are optional and default to label.
+  }}
+  <div class="field">
+    <div class="ui checkbox">
+      <input type="checkbox"
+             name="{{#if name}}{{name}}{{else}}{{label}}{{/if}}"
+             id="{{#if id}}{{id}}{{else}}{{label}}{{/if}}"
+             value="{{#if value}}{{value}}{{else}}{{label}}{{/if}}"
+             checked="{{checked}}">
+      <label>{{label}}</label>
+    </div>
+  </div>
+</template>
