+++ app/imports/ui/components/form-controls/checkbox-field.js
@@ -1,0 +1,5 @@
+import { Template } from 'meteor/templating';
+
+Template.Checkbox_Field.onRendered(function onRendered() {
+  this.$('.ui.checkbox').checkbox();
+});
