+++ app/imports/ui/components/form-controls/index.js
@@ -1,0 +1,12 @@
+import './checkbox-field.html';
+import './checkbox-field';
+import './checkbox-form-control.html';
+import './form-control-utilities';
+import './form-error-message.html';
+import './radio-field.html';
+import './radio-field';
+import './radio-form-control.html';
+import './select-form-control.html';
+import './select-form-control';
+import './text-form-control.html';
+import './textarea-form-control.html';
