+++ app/imports/ui/components/form-controls/radio-field.js
@@ -1,0 +1,5 @@
+import { Template } from 'meteor/templating';
+
+Template.Radio_Field.onRendered(function onRendered() {
+  this.$('.ui.radio.checkbox').checkbox();
+});
