+++ app/imports/ui/components/form-controls/select-form-control.js
@@ -1,0 +1,5 @@
+import { Template } from 'meteor/templating';
+
+Template.Select_Form_Control.onRendered(function onRendered() {
+  this.$('select.dropdown').dropdown();
+});
